Different activities and communication with server
The app contains three different views. In the first one the user
chooses a image, the language to translate from, the language to
translate to and if the translation should be enriched. The second view
shows the result of the computation from the server. In a third view,
the result text can be changed manually, and then sent to the server
again for another translation. The communication with the server is
handled by the class Enrich via Http Post method. If errors occur, the
status messages which are returned by the server are displayed.