Enrichment and design changes
First a new design was implemented. This includes a new color scheme
and a few changes regarding usage. The user can now edit the selected
image and then view it in a separate tab in the result and retry
activity. Further the translation can now be enriched by pressing long
on a word or phrase.
Readme, License and screenshots are also included.